+++ Sources/Player/Types/PlayerMetadata.swift
@@ -5,39 +5,7 @@
 //
 
 import AVFoundation
-import CoreMedia
 import MediaPlayer
-import PillarboxCore
-
-public struct ChapterMetadata: Equatable {
-    public let identifier: String?
-    public let title: String?
-    public let image: UIImage?
-    public let timeRange: CMTimeRange
-
-    fileprivate var timedNavigationMarker: AVTimedMetadataGroup {
-        .init(
-            items: [
-                .init(identifier: .commonIdentifierAssetIdentifier, value: identifier),
-                .init(identifier: .commonIdentifierTitle, value: title),
-                .init(identifier: .commonIdentifierArtwork, value: image?.pngData())
-            ].compactMap { $0 },
-            timeRange: timeRange
-        )
-    }
-
-    public init(
-        identifier: String? = nil,
-        title: String? = nil,
-        image: UIImage? = nil,
-        timeRange: CMTimeRange
-    ) {
-        self.identifier = identifier
-        self.title = title
-        self.image = image
-        self.timeRange = timeRange
-    }
-}
 
 public struct PlayerMetadata: Equatable {
     public static let empty = Self()
